refactor: use file_test_runner crate
