+++ src/graph.rs
@@ -5342,6 +5342,7 @@ mod tests {
     );
   }
 
+  #[cfg(feature = "fast_check")]
   #[tokio::test]
   async fn fast_check_dts() {
     let mut exports = IndexMap::new();
@@ -5409,6 +5410,7 @@ mod tests {
     assert!(dts.diagnostics.is_empty());
   }
 
+  #[cfg(feature = "fast_check")]
   #[tokio::test]
   async fn fast_check_external() {
     let mut exports = IndexMap::new();
